Add files via upload
This 87-page article has an introduction to the game, including a detailed description of the rules regarding the board and pieces. In particular, we describe the interesting way the pieces are set up initially, how they move, how they capture, and most importantly, how to win! Chess-like notation and easily readible board diagrams are used for examples and even a move-by-move presentation of two complete games.